backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
groq
python-multipart
python-dotenv
requests